Split dash components into separate folders
